refactor: move jpa repositories and models to infra package
