+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/port/factory/UserFactory.java
@@ -1,8 +1,8 @@
 package com.financialmonitoring.userservice.domain.port.factory;
 
 import com.financialmonitoring.userservice.adapter.dto.RegisterRequestDTO;
-import com.financialmonitoring.userservice.adapter.out.entity.Role;
-import com.financialmonitoring.userservice.adapter.out.entity.User;
+import com.financialmonitoring.userservice.infra.model.Role;
+import com.financialmonitoring.userservice.infra.model.User;
 
 import java.util.Set;
