+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/port/in/GetUserByEmailUseCase.java
@@ -1,6 +1,6 @@
 package com.financialmonitoring.userservice.domain.port.in;
 
-import com.financialmonitoring.userservice.adapter.out.entity.User;
+import com.financialmonitoring.userservice.infra.model.User;
 
 public interface GetUserByEmailUseCase {
     User getUserByEmail(String email);
