+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/service/AuthService.java
@@ -1,8 +1,8 @@
 package com.financialmonitoring.userservice.domain.service;
 
 import com.financialmonitoring.userservice.adapter.dto.*;
-import com.financialmonitoring.userservice.adapter.out.entity.Role;
-import com.financialmonitoring.userservice.adapter.out.entity.User;
+import com.financialmonitoring.userservice.infra.model.Role;
+import com.financialmonitoring.userservice.infra.model.User;
 import com.financialmonitoring.userservice.config.exception.BadRequestException;
 import com.financialmonitoring.userservice.domain.port.factory.UserFactory;
 import com.financialmonitoring.userservice.domain.port.in.GetUserByEmailUseCase;
