+++ user-service/src/test/java/com/financialmonitoring/userservice/domain/service/AuthServiceTest.java
@@ -4,8 +4,8 @@ import com.financialmonitoring.userservice.adapter.dto.LoginRequestDTO;
 import com.financialmonitoring.userservice.adapter.dto.LoginResponseDTO;
 import com.financialmonitoring.userservice.adapter.dto.RegisterRequestDTO;
 import com.financialmonitoring.userservice.adapter.dto.RegisterResponseDTO;
-import com.financialmonitoring.userservice.adapter.out.entity.Role;
-import com.financialmonitoring.userservice.adapter.out.entity.User;
+import com.financialmonitoring.userservice.infra.model.Role;
+import com.financialmonitoring.userservice.infra.model.User;
 import com.financialmonitoring.userservice.config.exception.BadRequestException;
 import com.financialmonitoring.userservice.domain.port.factory.UserFactory;
 import com.financialmonitoring.userservice.domain.port.out.AuthRepositoryPort;
